+++ genvalidity-bytestring/bench/Main.hs
@@ -7,7 +7,6 @@ module Main where
 import Criterion.Main as Criterion
 import Data.ByteString as Strict
 import Data.ByteString.Lazy as Lazy
-import Data.GenValidity
 import Data.GenValidity.ByteString ()
 import Data.GenValidity.Criterion
 
@@ -15,13 +14,13 @@ main :: IO ()
 main =
   Criterion.defaultMain
     [ bgroup
-        "Instances"
+        "generators"
         [ genValidBench @Strict.ByteString,
           genValidBench @Lazy.ByteString
         ],
       bgroup
-        "Approaches"
-        [ genBenchSizes "SB.pack <$> genValid" (Strict.pack <$> genValid),
-          genBenchSizes "LB.pack <$> genValid" (Lazy.pack <$> genValid)
+        "shrinkers"
+        [ shrinkValidBench @Strict.ByteString,
+          shrinkValidBench @Lazy.ByteString
         ]
     ]