+++ genvalidity-containers/src/Data/GenValidity/Map.hs
@@ -2,6 +2,7 @@
 
 module Data.GenValidity.Map
   ( genMapOf,
+    shrinkMapOf,
   )
 where
 
@@ -13,7 +14,10 @@ import Test.QuickCheck
 
 instance (Show k, Ord k, GenValid k, GenValid v) => GenValid (Map k v) where
   genValid = genMapOf genValid
-  shrinkValid = fmap M.fromList . shrinkValid . M.toList
+  shrinkValid = shrinkMapOf shrinkValid
 
 genMapOf :: Ord k => Gen (k, v) -> Gen (Map k v)
 genMapOf g = M.fromList <$> genListOf g
+
+shrinkMapOf :: Ord k => ((k, v) -> [(k, v)]) -> Map k v -> [Map k v]
+shrinkMapOf shrinker = fmap M.fromList . shrinkList shrinker . M.toList