+++ genvalidity-text/bench/Main.hs
@@ -7,25 +7,21 @@ module Main where
 import Criterion.Main as Criterion
 import Data.GenValidity
 import Data.GenValidity.Criterion
-import Data.GenValidity.Text
+import Data.GenValidity.Text ()
 import Data.Text as Strict
 import Data.Text.Lazy as Lazy
-import Test.QuickCheck
 
 main :: IO ()
 main =
   Criterion.defaultMain
     [ bgroup
-        "Instances"
-        [ genBenchSizes "Strict.Text" (genValid @Strict.Text),
-          genBenchSizes "Lazy.Text" (genValid @Lazy.Text)
+        "generators"
+        [ genBench "Strict.Text" (genValid @Strict.Text),
+          genBench "Lazy.Text" (genValid @Lazy.Text)
         ],
       bgroup
-        "Approaches"
-        [ genBenchSizes "via list (old version)" $ Strict.pack <$> genValid,
-          genBenchSizes "genText" genText,
-          genBenchSizes "genTextBy genValid" $ genTextBy genValid,
-          genBenchSizes "genTextBy (choose (minBound, maxBound)) (currently in use)" $
-            genTextBy (choose (minBound, maxBound))
+        "shrinkers"
+        [ shrinkBench "Strict.Text" (shrinkValid @Strict.Text),
+          shrinkBench "Lazy.Text" (shrinkValid @Lazy.Text)
         ]
     ]