+++ genvalidity-aeson/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.0.0] - 2022-08-30
+
+### Changed
+
+* Shrinking a `Value` now also shrinks to `Null` in every case.
+* Sped up generating `Key` by going via `Text` instead of going via `String`
+
 ## [1.0.0.1] - 2022-04-26
 
 ### Added