+++ genvalidity-case-insensitive/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [0.0.0.1] - 2022-08-30
+
 ## [0.0.0.0] - 2022-03-21
 
 Initial release: `GenValidity (CI a)`