+++ genvalidity-containers/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.0.0.1] - 2022-08-30
+
+### Added
+
+* `shrinkSetOf`, `shrinkSeqOf`, `shrinkMapOf`, and `shrinkTreeOf`
+
 ## [1.0.0.0] - 2021-11-20
 
 ### Changed