+++ genvalidity-time/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.0.0.1] - 2022-08-30
+
+### Changed
+
+* Sped up shrinking of `TimeOfDay`, `LocalTime` and `TimeOfDay`
+* Shrinking benchmarks
+
 ## [1.0.0.0] - 2021-11-20
 
 ### Changed